week9/requirements.txt
Flask
flasgger
python-dotenv
PyYAML